requirements.txt
annotated-types==0.7.0
audioread==3.1.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
decorator==5.2.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
language-tool-python==3.1.0
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markupsafe==3.0.3
matplotlib==3.10.8
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.15
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==24.0
platformdirs==4.5.1
pooch==1.8.2
preshed==3.0.12
psutil==7.1.3
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smart-open==7.5.0
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
textstat==0.7.12
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
tqdm==4.67.1
triton==3.5.1
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
xgboost==3.1.2
